IOTECO-18518: register plugins for android workmanager via reflections (#3)
This fixes the issue of background jobs running into an Exception when executing a function of a non registered plugin. This was only the case, when a background job was executed after the main app was killed.

Instead of registering the workmanager plugins within the apps MainActivity, we now use reflection to trigger the registration. As a simple & quick solution, the client implementing this plugin defines the following class: ch.belimo.belas.WorkmanagerPluginRegistrant and registers the workmanager plugins on the flutter engine within the static method registerWith(flutterEngine: FlutterEngine).